fix: reorganize mastra
