+++ examples/agent-frameworks/mastra/src/app/api/chat/route.ts
@@ -1,50 +1,9 @@
 import { MastraAgent } from "@ag-ui/mastra";
 import { Agent } from "@mastra/core/agent";
-import { createTool } from "@mastra/core/tools";
 import type { Message } from "@ag-ui/core";
 import { generateSystemPrompt } from "@openuidev/thesys-server";
 import { NextRequest } from "next/server";
-import { z } from "zod";
-
-// ========== Mock tools ==========
-const getWeather = createTool({
-  id: "get_weather",
-  description: "Get current weather for a city.",
-  inputSchema: z.object({ location: z.string().describe("City name") }),
-  execute: async ({ location }) => {
-    const knownTemps: Record<string, number> = {
-      tokyo: 22,
-      "san francisco": 18,
-      london: 14,
-      "new york": 25,
-      paris: 19,
-      sydney: 27,
-      mumbai: 33,
-      berlin: 16,
-    };
-    const temp = knownTemps[location.toLowerCase()] ?? Math.floor(Math.random() * 30 + 5);
-    return { location, temperature_celsius: temp, condition: "Clear" };
-  },
-});
-
-const getStockPrice = createTool({
-  id: "get_stock_price",
-  description: "Get current stock price for a ticker symbol.",
-  inputSchema: z.object({ symbol: z.string().describe("Ticker symbol, e.g. AAPL") }),
-  execute: async ({ symbol }) => {
-    const prices: Record<string, number> = {
-      AAPL: 189.84,
-      GOOGL: 141.8,
-      TSLA: 248.42,
-      MSFT: 378.91,
-      NVDA: 875.28,
-    };
-    const s = symbol.toUpperCase();
-    const price = prices[s] ?? Math.floor(Math.random() * 500 + 50);
-    return { symbol: s, price };
-  },
-});
-// ================================
+import { getStockPrice, getWeather } from "@/tools";
 
 const agent = new MastraAgent({
   agent: new Agent({
