+++ examples/agent-frameworks/mastra/src/app/layout.tsx
@@ -2,8 +2,8 @@ import type { Metadata } from "next";
 import "./globals.css";
 
 export const metadata: Metadata = {
-  title: "OpenUI Chat",
-  description: "Generative UI Chat with OpenAI SDK",
+  title: "OpenUI + Mastra Chat",
+  description: "Generative UI Chat with Mastra",
 };
 
 export default function RootLayout({
